Introduce a common interface for locating web elements
Previously, when looking for a nested element, client code had to
start from the `BrowserDriver` and then iterate a tree of
`WebElementDriver`s.

This change makes it easier to navigate a tree of elements by
introducing a common `WebElementLocator` interface shared by both
the `BrowserDriver` and the `WebElementDriver`. Selecting a nested
element can now be done with a simple for loop.

Note that `ElementLocator` has been renamed to `ElementFinder` to
avoid confusion with the new `WebElementLocator` interface.

Closes #17