Mouse over to element only when clicking with mouse
Previously `click`ing would first mouse over to element. This has
been removed and now we hover the mouse on the element only when
left clicking with mouse.

This solves and issues when using remote web driver with gecko, which
currently does not support mouse movement correctly. It also makes
senses to move the mouse only if clicking with the mouse.

See #16